+++ Editor/UI/Framework/AnalysisView.cs
@@ -113,7 +113,7 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
                 this);
 
             if (m_Desc.showDependencyView)
-                m_DependencyView = new DependencyView(new TreeViewState(), m_Desc.onDoubleClick);
+                m_DependencyView = new DependencyView(new TreeViewState(), m_Desc.onOpenIssue);
 
             if (m_TextFilter == null)
                 m_TextFilter = new TextFilter();
@@ -427,9 +427,16 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
 
         void SetRowsExpanded(bool expanded)
         {
-            var rows = m_Table.GetRows();
-            foreach (var row in rows)
-                m_Table.SetExpanded(row.id, expanded);
+            if (expanded)
+            {
+                var rows = m_Table.GetRows();
+
+                m_Table.SetExpanded(rows.Select(r => r.id).ToList());
+            }
+            else
+            {
+                m_Table.SetExpanded(new List<int>());
+            }
         }
 
         void Export(Func<ProjectIssue, bool> predicate = null)