+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+## [0.7.4-preview] - 2022-03-25
 * Added *OnRenderObject* and *OnWillRenderObject* to list of MonoBehavior critical contexts
 * Added Compilation Time property to Assemblies view
 * Added public API to get float/double custom property